more warning removals

+++ fqd_config.c
@@ -443,21 +443,21 @@ void fqd_config_http_stats(remote_client *client) {
     if(config->exchanges[i]) {
       fqd_exchange *e = config->exchanges[i];
       cprintf(client, " \"%.*s\": {\n", e->exchange.len, e->exchange.name);
-      cprintf(client, "  \"messages\": %llu,\n", e->stats->n_messages);
-      cprintf(client, "  \"octets\": %llu,\n", e->stats->n_bytes);
-      cprintf(client, "  \"no_route\": %llu,\n", e->stats->n_no_route);
-      cprintf(client, "  \"routed\": %llu,\n", e->stats->n_routed);
-      cprintf(client, "  \"dropped\": %llu\n", e->stats->n_dropped);
+      cprintf(client, "  \"messages\": %llu,\n", (long long unsigned int) e->stats->n_messages);
+      cprintf(client, "  \"octets\": %llu,\n", (long long unsigned int) e->stats->n_bytes);
+      cprintf(client, "  \"no_route\": %llu,\n", (long long unsigned int) e->stats->n_no_route);
+      cprintf(client, "  \"routed\": %llu,\n", (long long unsigned int) e->stats->n_routed);
+      cprintf(client, "  \"dropped\": %llu\n", (long long unsigned int) e->stats->n_dropped);
       cwrite(client, " },\n");
     }
   }
   cwrite(client, " \"_aggregate\": {\n");
-  cprintf(client, "  \"no_exchange\": %llu,\n", global_counters.n_no_exchange);
-  cprintf(client, "  \"messages\": %llu,\n", global_counters.n_messages);
-  cprintf(client, "  \"octets\": %llu,\n", global_counters.n_bytes);
-  cprintf(client, "  \"no_route\": %llu,\n", global_counters.n_no_route);
-  cprintf(client, "  \"routed\": %llu,\n", global_counters.n_routed);
-  cprintf(client, "  \"dropped\": %llu\n", global_counters.n_dropped);
+  cprintf(client, "  \"no_exchange\": %llu,\n", (long long unsigned int) global_counters.n_no_exchange);
+  cprintf(client, "  \"messages\": %llu,\n", (long long unsigned int) global_counters.n_messages);
+  cprintf(client, "  \"octets\": %llu,\n", (long long unsigned int) global_counters.n_bytes);
+  cprintf(client, "  \"no_route\": %llu,\n", (long long unsigned int) global_counters.n_no_route);
+  cprintf(client, "  \"routed\": %llu,\n", (long long unsigned int) global_counters.n_routed);
+  cprintf(client, "  \"dropped\": %llu\n", (long long unsigned int) global_counters.n_dropped);
   cwrite(client, " }\n");
   cwrite(client, "}\n");
   fqd_config_release(config);
